+++ src/components/Buttons.css
@@ -1,0 +1,16 @@
+.buttons {
+    width: 80%;
+    height: 50px;
+    position: fixed;
+    top: 80px;
+    display: flex;
+    flex-direction: row-reverse;
+    
+   
+}
+.buttons p {
+    padding: 20px;
+    cursor: pointer;
+    background-color: white;
+    box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
+}
