+++ src/components/Buttons.js
@@ -2,15 +2,17 @@ import React from 'react'
 import {useDispatch} from "react-redux";
 import {getCharactersAccion, siguienteCharacterAccion, 
     anteriorCharacterAccion} from "../APIS/MarvelAPI"
+import "./Buttons.css"
 
 
 const Buttons = (props) => {
     const dispatch = useDispatch();
     return (
         <div className="buttons" style={{ display: `${props.estado ? "flex" : "none"}`} }>
-            <button onClick={() => dispatch(getCharactersAccion())}>Get Personajes</button>
-            <button onClick={() => dispatch(anteriorCharacterAccion())}>Anterior</button> 
-            <button onClick={() => dispatch(siguienteCharacterAccion())}>Siguiente</button>
+            <p onClick={() => dispatch(getCharactersAccion())}>Get Personajes</p>
+            <p onClick={() => dispatch(siguienteCharacterAccion())}>Siguiente</p>
+            <p onClick={() => dispatch(anteriorCharacterAccion())}>Anterior</p> 
+            
       </div>
     )
 }
