+++ src/components/Galeria.css
@@ -1,11 +1,4 @@
-.buttons {
-    width: 80%;
-    height: 50px;
-    position: fixed;
-    top: 110px;
-    display: flex;
-    flex-direction: row-reverse;
-}
+
 .contenedor {
     display: flex;
     flex-wrap: wrap;
@@ -62,12 +55,14 @@
     height: 40px;
     margin-top: 20px;
     
+    
 }
 .starCard img {
     width: 30px;
     height: 30px;
     background-color: #b5a0a6;
     box-shadow: 1px 1px 2px black, 0 0 25px #00004a, 0 0 5px darkblue;
+    cursor: pointer;
 }
 
 @media (max-width: 700px){
