+++ src/components/WaitLoading.css
@@ -4,7 +4,7 @@
   top: 426px;
 
   display: block;
-  position: fixed;
+  position: absolute;
   justify-content: space-around;
   text-align: center;
 }
