+++ src/redux/store.js
@@ -1,7 +1,7 @@
 import {createStore, combineReducers, compose, applyMiddleware} from "redux";
 import thunk from "redux-thunk";
-import marvelReducer from "../APIS/MarvelReducer";
-import comicsReducer from "../APIS/ComicsReducer";
+import marvelReducer from "../APIS/MarvelAPI";
+import comicsReducer from "../APIS/ComicsAPI";
 
 const rootReducer = combineReducers({
     personajes: marvelReducer,
